% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{niche_percentile}
\alias{niche_percentile}
\title{Niche percentile shift}
\usage{
niche_percentile(psi_current, psi_future, occupied)
}
\arguments{
\item{psi_current}{Current squared niche distance for all cells.}

\item{psi_future}{Future squared niche distance for all cells.}

\item{occupied}{Current reference weights or indices used to define the
reference CDF.}
}
\value{
Data frame with current, future, and delta percentiles.
}
\description{
Niche percentile shift
}
