% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{niche_potential}
\alias{niche_potential}
\title{Squared niche distance}
\usage{
niche_potential(x, center, A)
}
\arguments{
\item{x}{Climate matrix in the fitted climate space.}

\item{center}{Current niche reference centre.}

\item{A}{Niche metric matrix.}
}
\value{
Numeric vector of squared distances from the current niche centre
in the fitted climatic space.
}
\description{
Squared niche distance
}
