% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aliases.R
\name{plot_climniche_class_summary}
\alias{plot_climniche_class_summary}
\title{Plot climniche class proportions}
\usage{
plot_climniche_class_summary(x, scope = c("current", "all"), title = NULL)
}
\arguments{
\item{x}{A fitted climniche object.}

\item{scope}{\code{"current"} for current occurrence/range cells or \code{"all"} for
all evaluated cells.}

\item{title}{Optional plot title.}
}
\value{
A ggplot object.
}
\description{
Plot climniche class proportions
}
