% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise.R
\name{plot_climniche_variable_contribution}
\alias{plot_climniche_variable_contribution}
\title{Plot mean climatic variable contribution}
\usage{
plot_climniche_variable_contribution(
  x,
  occupied_only = TRUE,
  variable_labels = NULL,
  title = NULL
)
}
\arguments{
\item{x}{A \code{climniche_fit} object.}

\item{occupied_only}{If TRUE, summarize occupied cells only.}

\item{variable_labels}{Optional named vector replacing variable labels.}

\item{title}{Optional plot title. Use \code{FALSE} to suppress it.}
}
\value{
A ggplot object.
}
\description{
Plot mean climatic variable contribution
}
