% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{write_climniche_report}
\alias{write_climniche_report}
\title{Write a climniche report to Markdown}
\usage{
write_climniche_report(report, file)
}
\arguments{
\item{report}{An object returned by \code{\link[=climniche_report]{climniche_report()}}.}

\item{file}{Output Markdown file.}
}
\value{
Invisibly returns \code{file}.
}
\description{
Write a climniche report to Markdown
}
