% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_report.R
\name{write_climniche_series_report}
\alias{write_climniche_series_report}
\title{Write a climniche time series report}
\usage{
write_climniche_series_report(report, file)
}
\arguments{
\item{report}{A \code{climniche_series_report} object.}

\item{file}{Output Markdown file.}
}
\value{
The file path, invisibly.
}
\description{
Write a climniche time series report
}
