% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{niche_potential}
\alias{niche_potential}
\title{Niche potential}
\usage{
niche_potential(x, center, A)
}
\arguments{
\item{x}{Standardized climate matrix.}

\item{center}{Realized niche center.}

\item{A}{Niche metric matrix.}
}
\value{
Numeric vector of quadratic niche displacement values.
}
\description{
Niche potential
}
