% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{niche_radius}
\alias{niche_radius}
\title{Niche radius}
\usage{
niche_radius(psi)
}
\arguments{
\item{psi}{Numeric niche-potential values.}
}
\value{
Numeric vector in sensitivity weighted climate distance units.
}
\description{
Niche radius
}
