% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{variable_contribution}
\alias{variable_contribution}
\title{Variable contribution to change in niche potential}
\usage{
variable_contribution(current, future, center, A)
}
\arguments{
\item{current}{Current standardized climate matrix.}

\item{future}{Future standardized climate matrix.}

\item{center}{Realised niche centre.}

\item{A}{Niche metric matrix.}
}
\value{
Matrix whose rows sum to the change in niche potential.
}
\description{
Variable contribution to change in niche potential
}
